requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
asttokens==3.0.0
asyncio==4.0.0
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
brotli==1.1.0
bs4==0.0.2
cachetools==6.2.0
certifi==2025.10.5
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
comm==0.2.3
courlan==1.3.2
cssselect==1.3.0
datasets==4.1.1
dateparser==1.2.2
debugpy==1.8.17
decorator==5.2.1
dill==0.4.0
distro==1.9.0
dotenv==0.9.9
executing==2.2.1
fastapi==0.118.2
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.9.0
google-auth==2.41.1
google-genai==1.41.0
goose3==3.1.20
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.10
hiredis==3.2.1
htmldate==1.6.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
ipykernel==6.30.1
ipython==9.6.0
ipywidgets==8.1.7
jedi==0.19.2
jieba3k==0.35.1
jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
justext==3.0.2
langchain-text-splitters==0.3.11
lxml==6.0.2
lxml-html-clean==0.4.3
markupsafe==3.0.3
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.5
newspaper3k==0.2.8
nltk==3.9.2
numpy==2.3.3
omegaconf==2.3.0
openai==2.2.0
orjson==3.11.3
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
prompt-toolkit==3.0.52
propcache==0.4.1
psutil==7.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyahocorasick==2.2.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.12.0
pydantic-core==2.41.1
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
readability-lxml==0.8.4.1
redis==6.4.0
regex==2025.9.18
replicate==1.0.7
requests==2.32.5
requests-file==2.1.0
requests-toolbelt==1.0.0
rsa==4.9.1
setuptools==80.9.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
starlette==0.48.0
sympy==1.14.0
tenacity==9.1.2
tinysegmenter==0.3
tld==0.13.1
tldextract==5.3.0
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
trafilatura==1.6.3
traitlets==5.14.3
triton==3.4.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.37.0
wcwidth==0.2.14
websockets==15.0.1
widgetsnbextension==4.0.14
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
